LayoutTests/fast/dom/HTMLMeterElement/meter-element-markup.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../resources/dump-as-markup.js"></script>
<style>
  meter { -webkit-appearance: none; }
</style>
</head>
<body>
  <div id="target">
    <meter value=70 max=100></meter>
    <meter min="0" max="10" low="3" high="6" optimum="5" value="10"></meter>
    <meter min="0" max="10" low="3" high="6" optimum="0" value="10"></meter>
  </div>
<script>
Markup.dump(document.getElementById("target"), "Both meter elements should have a nested shadow box with a width specified");
</script>
</body>
</html>
